requirements.txt
torch==1.10.1
torchvision==0.11.2
Pillow==8.4.0